+++ fotoobo/helpers/vault.py
@@ -25,7 +25,7 @@ class Client:  # pylint: disable=too-many-instance-attributes
     requests.
     """
 
-    def __init__(  # pylint: disable=too-many-arguments
+    def __init__(  # pylint: disable=too-many-arguments, too-many-positional-arguments
         self,
         url: str,
         namespace: str,