+++ tests/cli/fgt/test_cli_fgt_get.py
@@ -27,7 +27,7 @@ def test_cli_app_fgt_get_no_args() -> None:
     """Test fgt get with no arguments"""
     result = runner.invoke(app, ["-c", "tests/fotoobo.yaml", "fgt", "get"])
     assert result.exit_code == 0
-    assert "Usage: callback fgt get [OPTIONS] COMMAND" in result.stdout
+    assert "Usage: root fgt get [OPTIONS] COMMAND" in result.stdout
     assert "--help" in result.stdout
     assert "FortiGate get cmdb commands." in result.stdout