+++ WHATSNEW.md
@@ -1,6 +1,11 @@
 ### Added
 
+- Add CLI command fgt get cmdb firewall addrgrp
+- Add CLI command fgt get cmdb firewall address
+
 ### Changed
+
+- Optimize imports in CLI module
 
 ### Removed