+++ WHATSNEW.md
@@ -1,7 +1,10 @@
 ### Added
 
-- Add CLI command fgt get cmdb firewall addrgrp
+- Add generic api_get() method in tools/fgt/get
 - Add CLI command fgt get cmdb firewall address
+- Add CLI command fgt get cmdb firewall addrgrp
+- Add CLI command fgt get cmdb firewall service-custom
+- Add CLI command fgt get cmdb firewall service-group
 
 ### Changed