+++ WHATSNEW.md
@@ -1,5 +1,7 @@
 ### Added
 
+- Add the layers of fotoobo into the architecture documentation
+
 ### Changed
 
 ### Removed