+++ WHATSNEW.md
@@ -2,5 +2,7 @@
 
 ### Changed
 
+- Upgrade pygount to 1.8.0 due to import error
+
 ### Removed