+++ WHATSNEW.md
@@ -12,6 +12,7 @@
 - Fix some typing issues for Python3.8
 - Optimize imports in CLI module
 - Upgrade requests, jinja, pygount and virtualenv due to security issues and bugs
+- Upgrade all dependencies to be up to date
 
 ### Removed