+++ WHATSNEW.md
@@ -6,6 +6,7 @@
 - Add CLI command fgt get cmdb firewall service-custom
 - Add CLI command fgt get cmdb firewall service-group
 - Add the layers of fotoobo into the architecture documentation
+- Support for Python3.13
 
 ### Changed