+++ WHATSNEW.md
@@ -6,12 +6,14 @@
 - Add CLI command fgt get cmdb firewall service-custom
 - Add CLI command fgt get cmdb firewall service-group
 - Add the layers of fotoobo into the architecture documentation
+- Support for Python3.13
 
 ### Changed
 
 - Fix some typing issues for Python3.8
 - Optimize imports in CLI module
 - Upgrade requests, jinja, pygount and virtualenv due to security issues and bugs
+- Upgrade all dependencies to be up to date
 
 ### Removed