+++ WHATSNEW.md
@@ -10,6 +10,7 @@
 
 ### Changed
 
+- Fix syslog format: we do not have to set PRI_VAL as it is calculated by SysLogHandler
 - Fix some typing issues for Python3.8
 - Optimize imports in CLI module
 - Upgrade requests, jinja, pygount and virtualenv due to security issues and bugs